+++ src/app/components/ExploreOtherCareers.jsx
@@ -1,26 +1,14 @@
-import { Box } from '@chakra-ui/react';
 import CourseList from '@/app/components/CourseList';
 import cursos from '../content/cursosData';
+import shuffleArray from '@/app/utils/shuffled';
+import SectionHeading from './SectionHeading';
 
-function shuffleArray(array) {
-  const shuffled = [...array];
-  for (let i = shuffled.length - 1; i > 0; i--) {
-    const j = Math.floor(Math.random() * (i + 1));
-    [shuffled[i], shuffled[j]] = [shuffled[j], shuffled[i]];
-  }
-  return shuffled;
-}
-
-const ExploreOtherCareers = ({ currentCourseId, ishomePage, items }) => {
-  const quantity = items === undefined ? 3 : items;
-  const heading = ishomePage
-    ? 'Hecha un vistazo a nuestros cursos'
-    : 'Explora nuestras otras carreras';
+const ExploreOtherCareers = ({ currentCourseId, isHomePage, items, title }) => {
   let otherCourses = cursos.tecnicos.filter(
     (course) => course.id !== currentCourseId,
   );
 
-  if (ishomePage) {
+  if (isHomePage) {
     const cursosLibresMix = cursos.cursosLibresDiseno.concat(
       cursos.cursosLibresComputo,
     );
@@ -29,12 +17,13 @@ const ExploreOtherCareers = ({ currentCourseId, ishomePage, items }) => {
 
   const shuffledCourses = shuffleArray(otherCourses);
 
-  const limitedCourses = shuffledCourses.slice(0, quantity);
+  const limitedCourses = shuffledCourses.slice(0, items);
 
   return (
-    <Box>
-      <CourseList cursos={limitedCourses} showDescription title={heading} />
-    </Box>
+    <>
+      <SectionHeading title={title} />
+      <CourseList cursos={limitedCourses} showDescription />;
+    </>
   );
 };
