+++ src/app/content/cursosData.js
@@ -70,28 +70,25 @@ const cursos = {
       id: 'clDisenoPhotoShop',
       imgSrc: '/images/animacion-3d.jpg',
       title: 'PhotoShop',
-      details: {
-        modalidad: 'Cursos presenciales y virtuales',
-        edad: '15 años en adelante',
-        lugar: 'CETAV, Parque Libertad',
-        inversion: '₡25,000',
-        info: '8751-2802 e inscripción en el enlace',
-        infoLink: '',
-      },
+      details: [
+        'modalidad: Cursos presenciales y virtuales',
+        'edad: 15 años en adelante',
+        'lugar: CETAV, Parque Libertad',
+        'inversion: ₡25,000',
+        'info: 8751-2802 e inscripción en el enlace',
+      ],
     },
     {
       id: 'clDisenoIllustrador',
       imgSrc: '/images/animacion-3d.jpg',
       title: 'Illustrador',
-      url: 'https://bit.ly/Matr%C3%ADcula_Cursos_C%C3%B3mputo?fbclid=IwAR03fHSIYBcH01YfK9aX2TXrMJ-Pzfe5l6CAE3zpSXADAkDXDGb2lqgMaD8',
-      details: {
-        modalidad: 'Cursos presenciales y virtuales',
-        edad: '15 años en adelante',
-        lugar: 'CETAV, Parque Libertad',
-        inversion: '₡25,000',
-        info: '8751-2802 e inscripción en el enlace',
-        infoLink: '',
-      },
+      details: [
+        'modalidad: Cursos presenciales y virtuales',
+        'edad: 15 años en adelante',
+        'lugar: CETAV, Parque Libertad',
+        'inversion: ₡25,000',
+        'info: 8751-2802 e inscripción en el enlace',
+      ],
     },
     {
       id: 'clDisenoPremier',
