+++ src/app/page.js
@@ -1,11 +1,11 @@
 import { Box, Heading, Text } from '@chakra-ui/react';
 import InscriptionBanner from '@/app/components/InscriptionBanner';
 import LaboralFormation from '@/app/components/LaboralFormation';
-import images from './content/imageData';
+import images from '@/app/content/imageData';
 import CourseList from '@/app/components/CourseList';
-import cursos from './content/cursosData';
+import cursos from '@/app/content/cursosData';
 import OurAllies from '@/app/components/OurAllies';
-import ExploreOtherCareers from './components/ExploreOtherCareers';
+import ExploreOtherCareers from '@/app/components/ExploreOtherCareers';
 
 export default function Home() {
   return (
@@ -26,7 +26,11 @@ export default function Home() {
       <InscriptionBanner />
       <LaboralFormation />
       <OurAllies images={images} />
-      <ExploreOtherCareers items={2} ishomePage={true} />
+      <ExploreOtherCareers
+        isHomePage={true}
+        items={2}
+        title={'Hecha un vistazo a nuestros cursos'}
+      />
     </>
   );
 }
