+++ src/app/providers.jsx
@@ -2,7 +2,7 @@
 
 import { CacheProvider } from '@chakra-ui/next-js';
 import { ChakraProvider } from '@chakra-ui/react';
-import theme from '../theme';
+import theme from '@/theme';
 
 export function Providers({ children }) {
   return (
