+++ src/app/tecnicos/paid-media/page.js
@@ -28,7 +28,11 @@ export default function PaidMedia() {
       <TestimonyList testimonies={testimonies.web} />
       <SoftSkillsBanner />
       <OurTeachers teacherList={teachers.web} />
-      <ExploreOtherCareers currentCourseId={currentCourseId} />
+      <ExploreOtherCareers
+        currentCourseId={currentCourseId}
+        items={3}
+        title={'Explora nuestras carreras'}
+      />
       <FaqBanner />
     </>
   );
