+++ src/theme/index.js
@@ -1,6 +1,6 @@
 import { extendTheme } from '@chakra-ui/react';
-import typography from './foundations/typography';
-import colors from './foundations/colors';
+import typography from '@/theme/foundations/typography';
+import colors from '@/theme/foundations/colors';
 
 const overrides = {
   ...typography,
